#84 [refactor] API swagger 기준 네이밍 수정
