+++ src/api/attribute.ts
@@ -1,9 +1,9 @@
 import { api } from '@/api/client';
-import type { AttributeReturnDTO } from '@/api/types/attribute';
+import type { AttributesReturnDTO } from '@/api/types/attribute';
 import type { ApiResponse } from '@/api/types/api';
 
-export const getUserAbttributes = async (): Promise<AttributeReturnDTO[]> => {
+export const getUserAbttributes = async (): Promise<AttributesReturnDTO[]> => {
   const response =
-    await api.get<ApiResponse<AttributeReturnDTO[]>>('/attribute');
+    await api.get<ApiResponse<AttributesReturnDTO[]>>('/attribute');
   return response.data || [];
 };
