+++ src/types/attribute.ts
@@ -1,7 +1,7 @@
-import type { AttributeReturnDTO } from '@/api/types/attribute';
+import type { AttributesReturnDTO } from '@/api/types/attribute';
 import { ATTRIBUTE_TYPES } from '@/constants/attribute';
 
 export type AttributeType =
   (typeof ATTRIBUTE_TYPES)[keyof typeof ATTRIBUTE_TYPES];
 
-export type Attribute = AttributeReturnDTO;
+export type Attribute = AttributesReturnDTO;
